Start emulation

+++ src/apple/Shared/Metal/Atari800Renderer.h
@@ -2,7 +2,7 @@
 
 @interface Atari800Renderer : NSObject<MTKViewDelegate>
 
-@property (nonatomic, readonly) uint32_t * _Nullable screen;
+@property (nonatomic, readonly) uint8_t * _Nullable screen;
 
 - (nonnull instancetype)initWithMetalKitView:(nonnull MTKView *)view widthInPixels:(NSInteger)width heightInScanLines:(NSInteger)height;
